Start setting up HomeScreen with artemis.
